Re-authenticate after changing credentials
Hey,
not 100% sure about this but don't you have to re-authenticate after changing the oauth token? Once the function hits the limit, the second while loop randomly picks one of the credentials in `oauth_folder` until getting a credential with non-zero limit. If we end up with different credentials at the end of that process, we have to authenticate with the new token, no? My changes do that by a) adding `cr_old <- cr` before the loop to save the old credentials and b) adding a if condition after the loop that re-authenticate with the new token if `cr != cr_old`.

Best!